backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.3
sqlmodel==0.0.21
httpx==0.27.0
jinja2==3.1.4
python-multipart==0.0.9
pandas==2.2.2
numpy==2.0.0
scikit-learn==1.5.0
networkx==3.3
matplotlib==3.9.0
rich==13.7.1
pytest==8.2.0
pytest-asyncio==0.23.7
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
